requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu
pypdf